Updated to fix logical errors

+++ deleteEmptyFolders.py
@@ -1,6 +1,4 @@
 #!/usr/bin/env python3
-# Add checking of expiry of folder < 7 days old
-
 from sharefile.sharefile import Sharefile
 import os
 import sys
@@ -9,6 +7,7 @@ from dateutil import parser
 
 OUTFILE = f"{datetime.datetime.now():sharefile_log_%d_%m_%Y}"
 VERBOSE = True
+
 
 def main():
 	try:
@@ -21,30 +20,46 @@ def main():
 	delete_count = 0
 	reportUID = None
 	with open("/tmp/"+OUTFILE+"_"+sys.argv[1], "a") as f:
+		print("Found {} folders.".format(len(all_folders["value"])))
 		for i in all_folders["value"]:
 			if i["Name"] == "_Reports":
 				reportUID = i["Id"]
-				continue
-			if i["FileCount"] == 0:
-				dateDifference = (parser.parse(i['CreationDate']).replace(tzinfo=None) - datetime.datetime.now()).days
+				pass
+			if i["FileCount"] == 0 and i["Name"] != "_Reports":
+				dateDifference = (datetime.datetime.now() - parser.parse(i['CreationDate']).replace(tzinfo=None)).days
 				outString = "Found Empty Folder With Name: {}, ID: {}, Size: {}\n".format(i["Name"], i["Id"], i["FileCount"])
-				if dateDifference <= 15:
-					print(outString)
-					print("Skipping folder, created less than 15 days ago.")
-					continue
-				if VERBOSE:
-					print(outString)
-				f.write(outString)
-				if client.delete_item(i["Id"]):
-					print("Successfully Deleted")
-				delete_count += 1
+				if dateDifference > 14:
+					if VERBOSE:
+						print(outString)
+					f.write(outString)
+					try:
+						if client.delete_item(i["Id"]):
+							if VERBOSE:
+								print("Successfully Deleted")
+						delete_count += 1
+					except KeyboardInterrupt:
+						print("ctrl + c - Exiting")
+						exit()
+					except:
+						outstring = "Error deleting {} : {}".format(i["Id"], i["Name"])
+						if VERBOSE:
+							print(outstring)
+						f.write(outstring)
+				else:
+					if VERBOSE:
+						print("Skipping folder, created less than 15 days ago.")
 		if delete_count:
 			print("Deleted {} items.".format(delete_count))
 	print("Uploading log..")
 	client.upload_item(reportUID, "/tmp/"+OUTFILE+"_"+sys.argv[1], OUTFILE)
+
 if __name__ == "__main__":
 	if len(sys.argv) < 2:
-		print("This script removes all ")
-		print("USAGE:\n Environment Variables: SF_URL (ShareFile URL), SF_CID (Sharefile Client ID), SF_CS (Sharefile Client Secret), SF_USER (Sharefile service account username), SF_PWD (Sharefile user password)\n ./deleteEmptyFolders.py <parent folder ID item_id>")
+		print("This script removes all empty folders.")
+		print("USAGE:\n Environment Variables: \nSF_URL (ShareFile URL)\nSF_CID (Sharefile Client ID)\nSF_CS (Sharefile Client Secret)\nSF_USER (Sharefile service account username)\nSF_PWD (Sharefile user password)\n ./deleteEmptyFolders.py <parent folder ID item_id>")
 		exit()
-	main()
+	try:
+		main()
+	except KeyboardInterrupt:
+		print("ctrl + c - Exiting")
+		exit()
